LayoutTests/platform/ios/fast/webgpu/repro_285709-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'adapter.requestDevice')

WebGPU is only supported on Apple4 hardware or later or Mac2 hardware.
